+++ CHANGELOG.md
@@ -4,8 +4,10 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## dev
+
 * Bug
   * [cli] Fixing `--mount` option of `azk shell` command to comply with Docker's pattern (`local_folder:remote_folder`)
+  * [Docker] Fixing parse invalid instructions in build a Dockerfile, #391
   * [Tracking] Fixing timeout message when tracking an event;
 
 * Enhancements